Week 3. Alert

+++ components/FavoriteComponent.js
@@ -1,10 +1,13 @@
 import React, { Component } from 'react'
-import { FlatList, View, Text } from 'react-native'
-import { ListItem } from 'react-native-elements'
 import { connect } from 'react-redux'
-import { Loading } from './LoadingComponent'
-import { baseUrl } from '../shared/baseUrl'
+
+import { FlatList, View, Text, Alert } from 'react-native'
+import { ListItem } from 'react-native-elements'
 import Swipeout from 'react-native-swipeout'
+
+import { Loading } from './LoadingComponent'
+
+import { baseUrl } from '../shared/baseUrl'
 import { deleteFavorite } from '../redux/ActionCreators'
 
 const mapStateToProps = state => {
@@ -31,7 +34,24 @@ class Favorites extends Component {
 				{
 					text: 'Delete',
 					type: 'delete',
-					onPress: () => this.props.deleteFavorite(item.id)
+					onPress: () => {
+						Alert.alert(
+							'Delete Favorite?',
+							'Are you sure you wish to delete the favorite dish ' + item.name + '?',
+							[
+								{
+									text: 'Cancel',
+									onPress: () => console.log(item.name + 'Not Deleted'),
+									style: ' cancel'
+								},
+								{
+									text: 'OK',
+									onPress: () => this.props.deleteFavorite(item.id)
+								}
+							],
+							{ cancelable: false }
+						)
+					}
 				}
 			]
 			return (
